Huge refactor, switch to PWMOut compatible API.
This change reduces the scope of the PCA9685 driver to simply
providing PWMOut compatible objects for each channel. As a result,
the motor and servo can now be used with any PWM source, not just
the PCA9685. The new motor library is here:
https://github.com/adafruit/Adafruit_CircuitPython_Motor

Furthermore, this library is not aware of particular output configs
such as the specific Adafruit Motor FeatherWing. That knowledge has
been moved to a new FeatherWing library that handles all wing
related initialization. Its available here:
https://github.com/adafruit/Adafruit_CircuitPython_FeatherWing

Lastly, the new driver code and examples have been linted and
standard library files added.

Fixes #6 and fixes #7.