Correct PWM frequency prescale computation
Assuming this datasheet https://www.digikey.be/htmldatasheets/production/1640697/0/0/1/pca9685-datasheet.html#pf19 is correct, the prescale value computation is off by one. Prior to this change a PWM frequency of about 2 kHz could be set, and would be reported. After this change a maximum PWM frequency of 1526 Hz is accepted and reported, consistent with the datasheet.